+++ source/runner/_tests/_executor.test.ts
@@ -326,7 +326,7 @@ describe("setup", () => {
     await exec.handleResults(warnResults, dsl.git)
     expect(platform.updateStatus).toBeCalledWith(
       true,
-      ":warning: Danger found some issues. Don't worry, everything is fixable.",
+      "Found some issues. Don't worry, everything is fixable.",
       undefined,
       defaultConfig.dangerID
     )
@@ -342,7 +342,7 @@ describe("setup", () => {
     await exec.handleResults(failsResults, dsl.git)
     expect(platform.updateStatus).toBeCalledWith(
       false,
-      ":warning: Danger found some issues. Don't worry, everything is fixable.",
+      "Found some issues. Don't worry, everything is fixable.",
       undefined,
       defaultConfig.dangerID
     )